requirements.txt
asgiref==3.6.0
autopep8==2.0.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.21.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.4
Naked==0.1.32
ndg-httpsclient==0.5.1
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
plotly==5.13.1
proto-plus==1.22.2
protobuf==4.22.0
pushbullet.py==0.12.0
pushjack==1.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQRCode==1.2.1
python-magic==0.4.27
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rsa==4.9
shellescape==3.8.1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.2
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
websocket-client==1.5.1
XlsxWriter==3.0.8
